+++ broadcast-chan-conduit/CHANGELOG.md
@@ -1,3 +1,9 @@
+0.2.1 [2019.11.17]
+------------------
+* Fix resource management bug resulting from using `MonadUnliftIO` to run
+  `MonadResource` code in multiple threads. This version properly increments
+  resource count for every thread.
+
 0.2.0.2 [2019.03.30]
 --------------------
 * Update bounds for GHC 8.6